Remove fastdds files
