+++ include/fastrtps/rtps/builtin/discovery/participant/PDP.h
@@ -42,6 +42,7 @@ class RTPSReader;
 class WriterHistory;
 class ReaderHistory;
 class RTPSParticipantImpl;
+class RTPSParticipantListener;
 class BuiltinProtocols;
 class EDP;
 class TimedEvent;
@@ -347,6 +348,8 @@ protected:
     std::mutex temp_data_lock_;
     //!Participant data atomic access assurance
     std::recursive_mutex* mp_mutex;
+    //!To protect callbacks (ParticipantProxyData&)
+    std::mutex callback_mtx_;
 
     /**
      * Adds an entry to the collection of participant proxy information.
@@ -381,6 +384,21 @@ private:
 
     void check_remote_participant_liveliness(
             ParticipantProxyData* remote_participant);
+
+    void check_and_notify_type_discovery(
+            RTPSParticipantListener* listener,
+            const WriterProxyData& wdata) const;
+
+    void check_and_notify_type_discovery(
+            RTPSParticipantListener* listener,
+            const ReaderProxyData& rdata) const;
+
+    void check_and_notify_type_discovery(
+            RTPSParticipantListener* listener,
+            const string_255& topic_name,
+            const string_255& type_name,
+            const types::TypeIdentifier& type_id,
+            const types::TypeObject& type_obj) const;
 
     /**
      * Calculates the next announcement interval
